+++ example/http/server/async-ssl/http_server_async_ssl.cpp
@@ -390,7 +390,6 @@ class listener : public std::enable_shared_from_this<listener>
     net::io_context& ioc_;
     ssl::context& ctx_;
     tcp::acceptor acceptor_;
-    tcp::socket socket_;
     std::shared_ptr<std::string const> doc_root_;
 
 public:
@@ -402,7 +401,6 @@ public:
         : ioc_(ioc)
         , ctx_(ctx)
         , acceptor_(ioc)
-        , socket_(beast::make_strand(ioc))
         , doc_root_(doc_root)
     {
         beast::error_code ec;
@@ -453,15 +451,16 @@ public:
     void
     do_accept()
     {
+        // The new connection gets its own strand
         acceptor_.async_accept(
-            socket_,
+            beast::make_strand(ioc_),
             beast::bind_front_handler(
                 &listener::on_accept,
                 shared_from_this()));
     }
 
     void
-    on_accept(beast::error_code ec)
+    on_accept(beast::error_code ec, tcp::socket socket)
     {
         if(ec)
         {
@@ -471,13 +470,10 @@ public:
         {
             // Create the session and run it
             std::make_shared<session>(
-                std::move(socket_),
+                std::move(socket),
                 ctx_,
                 doc_root_)->run();
         }
-
-        // Make sure each session gets its own strand
-        socket_ = tcp::socket(beast::make_strand(ioc_));
 
         // Accept another connection
         do_accept();