+++ example/websocket/server/fast/websocket_server_fast.cpp
@@ -243,7 +243,6 @@ class async_listener : public std::enable_shared_from_this<async_listener>
 {
     net::io_context& ioc_;
     tcp::acceptor acceptor_;
-    tcp::socket socket_;
 
 public:
     async_listener(
@@ -251,7 +250,6 @@ public:
         tcp::endpoint endpoint)
         : ioc_(ioc)
         , acceptor_(beast::make_strand(ioc))
-        , socket_(beast::make_strand(ioc))
     {
         beast::error_code ec;
 
@@ -301,15 +299,16 @@ public:
     void
     do_accept()
     {
+        // The new connection gets its own strand
         acceptor_.async_accept(
-            socket_,
+            beast::make_strand(ioc_),
             beast::bind_front_handler(
                 &async_listener::on_accept,
                 shared_from_this()));
     }
 
     void
-    on_accept(beast::error_code ec)
+    on_accept(beast::error_code ec, tcp::socket socket)
     {
         if(ec)
         {
@@ -318,11 +317,8 @@ public:
         else
         {
             // Create the async_session and run it
-            std::make_shared<async_session>(std::move(socket_))->run();
+            std::make_shared<async_session>(std::move(socket))->run();
         }
-
-        // Make sure each session gets its own strand
-        socket_ = tcp::socket(beast::make_strand(ioc_));
 
         // Accept another connection
         do_accept();