+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 217:
 
 * websocket idle pings
 * RatePolicy documentation
+* Pass strand to async_accept
 
 --------------------------------------------------------------------------------